backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio>=5.5.0
eventlet>=0.37.0
ultralytics>=8.3.86
opencv-python>=4.8.0
Pillow>=11.1.0
numpy>=1.23.0
# gspread and google-auth removed -- using SQLite instead
python-dotenv==1.0.1
# Machine Learning Dependencies
scikit-learn>=1.3.0
pandas>=1.5.0
joblib>=1.3.0
